// packages/wizzi.site/dist/src/server/features/config/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi-js\dist\lib\artifacts\ts\module\gen\main.js
    package: wizzi-js@0.7.8
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi.site\.wizzi\server\src\features\config\types.ts.ittf
    utc time: Wed, 09 Jun 2021 05:04:16 GMT
*/
export type ConfigType = { 
    port: number;
    sessionSecret: string;
    noCache: boolean;
    corsClientOrigin: string;
    mongoPath: string;
    mongoUser: string;
    mongoPassword: string;
    auth0Domain: string;
    auth0PackiClientId: string;
    auth0PackiClientSecret: string;
    auth0PackiCallbackUrl: string;
    auth0PackiBackendAppId: string;
    auth0PackiBackendAppSecret: string;
    auth0PackiApiId: string;
    auth0IssuerBaseURL: string;
    auth0ClientId: string;
    auth0BaseURL: string;
    auth0Secret: string;
    githubClientId: string;
    githubClientSecret: string;
    githubCallbackURL: string;
    isWizziDev: boolean;
    ittfPath: string;
    dataPath: string;
    metaHtmlIttfPath: string;
    metaFolderIttfPath: string;
    metaHtmlTextPath: string;
    packiStoreFolder: string;
    packiTemplatesFolder: string;
    packiApiEndpoint: string;
};